Elaborate more on how to fill in the extensions
There has been some debate on the list on how much of RFC 5280, a
profile of X.509, applies to unsigned certificates.

Elaborate on this a bit more, both to better capture the intent and
formally prescribe the cases where applicability is unclear.

It was pointed out on the list that RFC 5280 explicitly mandates the AKI
extension. The AKI extension doesn't make sense here and is very
unlikely to pose any interoperability concerns, but it is formally at
best unclear whether it applies. To avoid that ambiguity, I've made
the text explicitly update that requirement. That means now this draft
needs to update 5280, so I've gone ahead and done that.

In writing this, I also realized that there's some ambiguity around key
usage and basic constraints. Or rather, we've now fixed that ambiguity,
but should probably say so explicitly.

This does rather add to the scope of this section at WGLC. But I think
this is appropriate:

1. The comments on the list suggest people find the lack of guidance
   confusing.

2. All the way back to when the document was first adopted, the
   introduction said, "This document defines a profile for unsigned
   X.509 certificates", so defining how to issue them in full seems in
   scope.

(There was another comment on the list about the OID to use in the
placeholder issuer. This change is not intended to address that. I've
left it alone for now pending guidance on where to allocate the
replacement OID.)